build(deps): upgrade @angular-eslint/* BREAKING CHANGE: Requires angular 19.
