+++ packages/eslint-plugin-rules/package.json
@@ -27,11 +27,11 @@
   },
   "dependencies": {
     "@eslint/eslintrc": "^2.1.4",
-    "@typescript-eslint/utils": "^7.16.0"
+    "@typescript-eslint/utils": "^8.18.0"
   },
   "devDependencies": {
-    "@typescript-eslint/parser": "^7.16.0",
-    "@typescript-eslint/rule-tester": "^7.16.0"
+    "@typescript-eslint/parser": "^8.18.0",
+    "@typescript-eslint/rule-tester": "^8.18.0"
   },
   "peerDependencies": {
     "eslint": "^8.56.0"
