+++ packages/eslint-plugin-rules/src/rules/deny-parent-index-file-import.spec.ts
@@ -1,36 +1,24 @@
-import { RuleTester } from '@typescript-eslint/rule-tester'
-import { join } from 'node:path'
+import { createRuleTester } from '../utils/rule-tester'
 import { DenyParentIndexFileImportMessageIds, denyParentIndexFileImportRule } from './deny-parent-index-file-import'
-const tsRootDirectory = join(__dirname, '../..', 'test')
-const ruleTester = new RuleTester({
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 2022,
-    tsconfigRootDir: tsRootDirectory,
-    project: './tsconfig.json',
-  },
-})
+
+const ruleTester = createRuleTester()
 
 ruleTester.run('deny-parent-index-file-import', denyParentIndexFileImportRule, {
   valid: [
     {
-      // language=typescript
       code: `import { foo } from '@angular/core'`,
     },
   ],
   invalid: [
     {
-      // language=typescript
       code: `import { foo } from '..'`,
       errors: [{ messageId: DenyParentIndexFileImportMessageIds.DENIED_IMPORT }],
     },
     {
-      // language=typescript
       code: `import { foo } from '../..'`,
       errors: [{ messageId: DenyParentIndexFileImportMessageIds.DENIED_IMPORT }],
     },
     {
-      // language=typescript
       code: `import { foo } from 'aws-sdk'
       import { MyService } from '../..'`,
       errors: [{ messageId: DenyParentIndexFileImportMessageIds.DENIED_IMPORT, data: { path: '../..' } }],
